# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A high-performance REST API service for scraping Google search results using SerpAPI. Designed for lead generation workflows.",
        "title": "Lead Gen Worker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/reports": {
            "get": {
                "description": "Retrieves comprehensive usage reports including token usage, costs, and lead generation metrics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get usage reports",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to get reports for",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date for the report period (RFC3339 format)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date for the report period (RFC3339 format)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usage reports",
                        "schema": {
                            "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.ReportsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reports/daily": {
            "get": {
                "description": "Retrieves usage statistics aggregated by day for charts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get daily usage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339 or YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339 or YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daily usage statistics",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.DailyUsage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reports/operations": {
            "get": {
                "description": "Retrieves usage statistics grouped by AI operation type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get operation statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339 or YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339 or YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation statistics",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.OperationStats"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reports/summary": {
            "get": {
                "description": "Retrieves a quick summary of token usage and costs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get usage summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to get summary for",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339 or YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339 or YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usage summary",
                        "schema": {
                            "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.UsageSummary"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "webstar_noturno-leadgen-worker_internal_dto.DailyUsage": {
            "description": "Usage statistics aggregated by day",
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "failed_calls": {
                    "type": "integer"
                },
                "successful_calls": {
                    "type": "integer"
                },
                "total_calls": {
                    "type": "integer"
                },
                "total_cost_usd": {
                    "type": "number"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
...
        "webstar_noturno-leadgen-worker_internal_dto.LeadGenerationStats": {
            "description": "Statistics specific to lead generation pipeline",
            "type": "object",
            "properties": {
                "avg_cost_per_lead": {
                    "type": "number"
                },
                "avg_leads_per_job": {
                    "type": "number"
                },
                "total_emails_generated": {
                    "type": "integer"
                },
                "total_jobs_processed": {
                    "type": "integer"
                },
                "total_leads_generated": {
                    "type": "integer"
                },
                "total_reports_generated": {
                    "type": "integer"
                }
            }
        },
        "webstar_noturno-leadgen-worker_internal_dto.ModelUsage": {
            "description": "Usage statistics by AI model",
            "type": "object",
            "properties": {
                "avg_tokens_per_call": {
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "total_calls": {
                    "type": "integer"
                },
                "total_cost_usd": {
                    "type": "number"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
        "webstar_noturno-leadgen-worker_internal_dto.OperationStats": {
            "description": "Statistics for a specific AI operation type",
            "type": "object",
            "properties": {
                "avg_duration_ms": {
                    "type": "number"
                },
                "failed_calls": {
                    "type": "integer"
                },
                "operation_type": {
                    "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.OperationType"
                },
                "success_rate": {
                    "type": "number"
                },
                "successful_calls": {
                    "type": "integer"
                },
                "total_calls": {
                    "type": "integer"
                },
                "total_cost_usd": {
                    "type": "number"
                },
                "total_input_tokens": {
                    "type": "integer"
                },
                "total_output_tokens": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
        "webstar_noturno-leadgen-worker_internal_dto.OperationType": {
            "type": "string",
            "enum": [
                "data_extraction",
                "pre_call_report",
                "cold_email",
                "website_scraping"
            ],
            "x-enum-varnames": [
                "OperationDataExtraction",
                "OperationPreCallReport",
                "OperationColdEmail",
                "OperationWebsiteScraping"
            ]
        },
        "webstar_noturno-leadgen-worker_internal_dto.ReportPeriod": {
            "description": "Time range covered by the report",
            "type": "object",
            "properties": {
                "days_count": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "webstar_noturno-leadgen-worker_internal_dto.ReportsResponse": {
            "description": "Complete reports response for dashboard visualization",
            "type": "object",
            "properties": {
                "by_model": {
                    "description": "ByModel contains stats grouped by AI model",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.ModelUsage"
                    }
                },
                "by_operation": {
                    "description": "ByOperation contains stats grouped by operation type",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.OperationStats"
                    }
                },
                "daily_usage": {
                    "description": "DailyUsage contains usage over time",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.DailyUsage"
                    }
                },
                "lead_generation": {
                    "description": "LeadGeneration contains lead-specific metrics",
                    "allOf": [
                        {
                            "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.LeadGenerationStats"
                        }
                    ]
                },
                "period": {
                    "description": "Period indicates the date range of the report",
                    "allOf": [
                        {
                            "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.ReportPeriod"
                        }
                    ]
                },
                "summary": {
                    "description": "Summary contains overall usage metrics",
                    "allOf": [
                        {
                            "$ref": "#/definitions/webstar_noturno-leadgen-worker_internal_dto.UsageSummary"
                        }
                    ]
                }
            }
        },
...
        "webstar_noturno-leadgen-worker_internal_dto.UsageSummary": {
            "description": "Overall usage summary with key metrics",
            "type": "object",
            "properties": {
                "avg_cost_per_call": {
                    "type": "number"
                },
                "avg_duration_ms": {
                    "type": "number"
                },
                "avg_tokens_per_call": {
                    "type": "number"
                },
                "failed_calls": {
                    "type": "integer"
                },
                "success_rate": {
                    "type": "number"
                },
                "successful_calls": {
                    "type": "integer"
                },
                "total_calls": {
                    "type": "integer"
                },
                "total_cost_usd": {
                    "type": "number"
                },
                "total_input_tokens": {
                    "type": "integer"
                },
                "total_output_tokens": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
...
    }
}